requirements.txt
# Web UI
streamlit>=1.25.0
requests==2.28.2

# API
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7

# Blockchain
web3>=6.0.0
python-dotenv==1.0.0

# Explicitly set protobuf version to resolve conflict
protobuf>=4.21.6

# Cryptography
cryptography==40.0.2
merkletools==1.0.3
pysha3==1.0.2
# pygroupsig is already included in the project directory

# IPFS
ipfshttpclient==0.8.0a2

# cdp-sdk is optional and not required for basic functionality